+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
+
+
+## [v0.47.0]
+Release date: 2026-05-19
+
 - Fix `entropy_mutual` (mutual information) when the input state is a pure state (`Ket`). ([#716])
 - Use `expm1` in `n_thermal` for stability. ([#717], [#718])
 - Use `CUDACore` and `cuSPARSE` as dependencies instead of `CUDA`. ([#687])
@@ -362,6 +367,7 @@ Release date: 2024-11-13
 [v0.44.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.44.0
 [v0.45.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.45.0
 [v0.46.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.46.0
+[v0.47.0]: https://github.com/qutip/QuantumToolbox.jl/releases/tag/v0.47.0
 [#86]: https://github.com/qutip/QuantumToolbox.jl/issues/86
 [#139]: https://github.com/qutip/QuantumToolbox.jl/issues/139
 [#271]: https://github.com/qutip/QuantumToolbox.jl/issues/271